Logs configuration improvements
