+++ src/bootstrap/bootstrap.py
@@ -4,8 +4,9 @@ from dependency_injector.providers import Object
 from domains import init_celery, init_domains
 from pydantic import BaseModel, ConfigDict
 
-from .config import AppConfig, init_logger
+from .config import AppConfig
 from .di_container import Container
+from .logs import init_logger
 from .storage import init_storage
